Reduce the amount of string concatenation.
Previously, the remaining buffer and the newly received data were
concatenated prior to splitting.
However, we know for sure that the remaining buffer is free of
seperators. That means we can split the newly received data first,
and then paste the remaining buffer in front of the first so
obtained line.